Split out Windows-only tests
